Cleaning up reporting w/ TestResult
