+++ tests/runner/src/tests/producer_stress.rs
@@ -9,8 +9,8 @@ use fluvio_test_util::test_meta::derive_attr::TestRequirements;
 use fluvio_test_util::test_meta::environment::EnvironmentSetup;
 use fluvio_test_util::test_meta::{TestOption, TestCase};
 use fluvio_test_util::test_meta::test_result::TestResult;
-use fluvio_test_util::test_runner::FluvioTestDriver;
-use fluvio_test_util::test_runner::fluvio_test_meta::FluvioTestMeta;
+use fluvio_test_util::test_runner::test_driver::FluvioTestDriver;
+use fluvio_test_util::test_runner::test_meta::FluvioTestMeta;
 use async_lock::RwLock;
 
 #[derive(Debug, Clone)]
