+++ tests/runner/src/utils/test_meta/test_result.rs
@@ -2,8 +2,9 @@ use std::any::Any;
 use std::fmt::{self, Debug, Display, Formatter};
 use prettytable::{table, row, cell};
 use std::time::Duration;
+use hdrhistogram::Histogram;
 
-#[derive(Debug, Clone, Default)]
+#[derive(Debug, Clone)]
 pub struct TestResult {
     pub success: bool,
     pub duration: Duration,
@@ -16,8 +17,36 @@ pub struct TestResult {
     pub num_consumers: u64,
     pub num_topics: u64,
     pub topic_create_latency: u64,
+
+    pub topic_create_latency_histogram: Histogram<u64>,
+    pub produce_latency_histogram: Histogram<u64>,
+    pub consume_latency_histogram: Histogram<u64>,
 }
 
+impl Default for TestResult {
+    fn default() -> Self {
+        Self {
+            success: false,
+            duration: Duration::new(0, 0),
+
+            num_producers: 0,
+            num_consumers: 0,
+            num_topics: 0,
+
+            bytes_produced: 0,
+            bytes_consumed: 0,
+
+            produce_latency: 0,
+            consume_latency: 0,
+            topic_create_latency: 0,
+
+            produce_latency_histogram: Histogram::<u64>::new_with_bounds(1, u64::MAX, 2).unwrap(),
+            consume_latency_histogram: Histogram::<u64>::new_with_bounds(1, u64::MAX, 2).unwrap(),
+            topic_create_latency_histogram: Histogram::<u64>::new_with_bounds(1, u64::MAX, 2)
+                .unwrap(),
+        }
+    }
+}
 impl TestResult {
     pub fn as_any(&self) -> &dyn Any {
         self
@@ -38,32 +67,98 @@ impl TestResult {
     }
 }
 
-// TODO: Parse the time scalars into Duration
 impl Display for TestResult {
     fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
         let success_str = format!("{}", self.success);
         let duration_str = format!("{:?}", self.duration);
-        let producer_latency_str = format!("{:?}", Duration::from_nanos(self.produce_latency));
-        let consumer_latency_str = format!("{:?}", Duration::from_nanos(self.consume_latency));
-        let topic_create_latency_str =
-            format!("{:?}", Duration::from_nanos(self.topic_create_latency));
 
-        let table = table!(
+        let basic_results_table = table!(
             [b->"Test Results"],
             ["Pass?", b->success_str],
-            ["Duration", duration_str],
-            //
-
-            ["# topics created", self.num_topics],
-            ["topic create latency 99%", topic_create_latency_str],
-            ["# producers created", self.num_producers],
-            ["bytes produced", self.bytes_produced],
-            ["producer latency 99%", producer_latency_str],
-            ["# consumers created", self.num_consumers],
-            ["bytes consumed", self.bytes_consumed],
-            ["consumer latency 99%", consumer_latency_str]
+            ["Duration", duration_str]
         );
 
-        write!(f, "{}", table)
+        let topic_create_latency_avg = format!(
+            "{:.2?}",
+            Duration::from_nanos(self.topic_create_latency_histogram.mean() as u64)
+        );
+        let topic_create_latency_p50 = format!(
+            "{:.2?}",
+            Duration::from_nanos(
+                self.topic_create_latency_histogram
+                    .value_at_percentile(50.0)
+            )
+        );
+        let topic_create_latency_p99 = format!(
+            "{:.2?}",
+            Duration::from_nanos(
+                self.topic_create_latency_histogram
+                    .value_at_percentile(99.0)
+            )
+        );
+        let topic_create_latency_p999 = format!(
+            "{:.2?}",
+            Duration::from_nanos(
+                self.topic_create_latency_histogram
+                    .value_at_percentile(99.9)
+            )
+        );
+
+        let producer_latency_avg = format!(
+            "{:.2?}",
+            Duration::from_nanos(self.produce_latency_histogram.mean() as u64)
+        );
+        let producer_latency_p50 = format!(
+            "{:.2?}",
+            Duration::from_nanos(self.produce_latency_histogram.value_at_percentile(50.0))
+        );
+        let producer_latency_p99 = format!(
+            "{:.2?}",
+            Duration::from_nanos(self.produce_latency_histogram.value_at_percentile(99.0))
+        );
+        let producer_latency_p999 = format!(
+            "{:.2?}",
+            Duration::from_nanos(self.produce_latency_histogram.value_at_percentile(99.9))
+        );
+
+        let consumer_latency_avg = format!(
+            "{:.2?}",
+            Duration::from_nanos(self.consume_latency_histogram.mean() as u64)
+        );
+        let consumer_latency_p50 = format!(
+            "{:.2?}",
+            Duration::from_nanos(self.consume_latency_histogram.value_at_percentile(50.0))
+        );
+        let consumer_latency_p99 = format!(
+            "{:.2?}",
+            Duration::from_nanos(self.consume_latency_histogram.value_at_percentile(99.0))
+        );
+        let consumer_latency_p999 = format!(
+            "{:.2?}",
+            Duration::from_nanos(self.consume_latency_histogram.value_at_percentile(99.9))
+        );
+
+        let perf_results_table = table!(
+            [b->"Perf Results"],
+
+            [b->"Created", "#"],
+            ["Topic", self.num_topics],
+            ["Producer", self.num_producers],
+            ["Consumer", self.num_consumers],
+
+            [b->""],
+            [b->"Throughput", b->"Bytes"],
+            ["Producer", self.bytes_produced],
+            ["Consumer", self.bytes_consumed],
+
+            [b->""],
+            [b->"Latency", b->"Average", b->"P50", b->"P99", b->"P999"],
+            ["Topic create", topic_create_latency_avg, topic_create_latency_p50, topic_create_latency_p99, topic_create_latency_p999],
+            ["Producer", producer_latency_avg, producer_latency_p50, producer_latency_p99, producer_latency_p999],
+            ["Consumer", consumer_latency_avg, consumer_latency_p50, consumer_latency_p99, consumer_latency_p999]
+        );
+
+        write!(f, "{}", basic_results_table)?;
+        write!(f, "\n{}", perf_results_table)
     }
 }
